Add attic nix cache server to sean-pi-public host
Deploy an attic server for public deployment at nix-cache.mistersg.net
and privately at nix-cache-local.mistersg.net.

Previously, the attic module was used for testing and debugging the
attic server locally, but now this is being repurposed as a production
grade deployment. This will allow me to host a public nix cache
without having to deal with the limitations of the cachix
free tier.