Add Thunderbird to i3 launchers
Thunderbird is now used daily to sync email and calendar events. This
works better than webmail, since it allows me to merge my work and
caldav calendars.

Update the Telegram launcher to use the correct executable (Telegram
insead of telegram-desktop). This was changed by some release in the
last year or so