Fix broken nixpkgs open-webui package
The open-webui package is currently missing required dependency
`itsdangerous`. Override the derivations to add it.

Add searxng service. SearXNG is a distributed privacy-focus search
engine aggregator. It is useful for as an AI tool, but I may set it as
my default search engine too.

Add oterm package. oterm is a TUI for running local LLMs. It's simpler
than Open WebUI, and makes integrating MCP servers a lot easier.